requirements.txt
Flask==3.1.1
python-dotenv
Flask-Mail
requests